minor refactoring in db layer
