+++ weather_dl_v2/fastapi-server/database/license_handler.py
@@ -43,14 +43,6 @@ class LicenseHandler(abc.ABC):
         pass
 
     @abc.abstractmethod
-    def _create_license_queue(self, license_id: str, client_name: str) -> None:
-        pass
-
-    @abc.abstractmethod
-    def _remove_license_queue(self, license_id: str) -> None:
-        pass
-
-    @abc.abstractmethod
     def _get_license_without_deployment(self) -> list:
         pass
 
@@ -106,12 +98,6 @@ class LicenseHandlerMock(LicenseHandler):
             }
         ]
 
-    def _create_license_queue(self, license_id: str, client_name: str) -> None:
-        logger.info("Added L1 queue in 'queues' collection. Update_time: 00000.")
-
-    def _remove_license_queue(self, license_id: str) -> None:
-        logger.info("Removed L1 queue in 'queues' collection. Update_time: 00000.")
-
     def _get_license_without_deployment(self) -> list:
         return []
 
@@ -159,16 +145,6 @@ class LicenseHandlerFirestore(LicenseHandler):
         for snapshot in snapshot_list:
             result.append(self.db.collection(self.collection).document(snapshot.id).get().to_dict())
         return result
-
-    def _create_license_queue(self, license_id: str, client_name: str) -> None:
-        result: WriteResult = self.db.collection('queues').document(license_id).set(
-            {"license_id": license_id, "client_name": client_name, "queue": []}
-        )
-        logger.info(f"Added {license_id} queue in 'queues' collection. Update_time: {result.update_time}.")
-
-    def _remove_license_queue(self, license_id: str) -> None:
-        timestamp = self.db.collection('queues').document(license_id).delete()
-        logger.info(f"Removed {license_id} queue in 'queues' collection. Update_time: {timestamp}.")
 
     def _get_license_without_deployment(self) -> list:
         snapshot_list = self.db.collection(self.collection).where('k8s_deployment_id', '==', '').get()
